Fix incrementality (soundness) bug.
Backends no longer directly depend on frontends. Because previously we said they did, they were run immediately after a frontend task they depend on in bottomup fashion, before the main task is even run. The main task actually takes care of passing the new output of the frontend task to the backend task as a new input. But PIE would not run the backend task again with a different input and the same key. The same key is good to have so we don't generate lots of new tasks that are basically about the same thing (like stack vs heap based memory management).

This all was the way it was because frontends used to generate files that the backends would read. Therefore there was a dependency, you really needed the frontends to be run before the backends. Of course this was implicitly done by the main task, but PIE required that this was acknowledged in the dependency system. In case you "required" the tasks in the wrong order.

Now we're actually using the implicit dependency that the main task maintains between front- and backends. By not using them explicitly (e.g. by depending on part of the output of the frontend tasks inside the backend task thru a require) we put less pressure on PIE because it needs to take care of less dependencies. Simple ones like these can be clearly modelled in the code instead of the dynamic system.