Use absolute paths instead of relative in str incr
Windows filesystems can have multiple roots (like C: and E:). If Spoofax
is installed in one, and the Spoofax project resides in another, then
libraries shipped with Spoofax will have different roots for the path.
Therefore we cannot, in general, work with relative paths from the
project root. On linux this worked because it's all the same root. But
there the relative paths for libraries shipped with Spoofax were also a
bit ugly with all the ../../.. at the start.