A couple of days work:
Started building the new incremental compiler next to the old one, copying over bits and pieces.
This version is based on a new architecture, trying out less caching and much more direct dependencies on tasks instead of copying outputs of tasks to inputs to other tasks. With plenty of function-based stampers to select only the parts of the output we're interested in, this should make the Pie part much faster. Also reimplementing some simple Stratego things in Java in the hopes of removing some list-creating/list-merging overhead. Finally the plan is to also overhaul the bits of static analysis still in Java to make that cheaper (not in this commit).

Done so far: Service interface for resolving imports and getting asts, Front for splitting the ast into ModuleData, Resolve for walking the import graph and indexing all ModuleData into a GlobalData.
TODO: Last bit of Front, Analysis for Java based static analysis and type system calls, Compile for calling Analysis and then running Back tasks, Back for each definition.