Incremental Compiler v2 (#15)
This is the new incremental compiler design for Stratego. The general idea is still the same: 1 Front task per file, 1 Back task per strategy (generating one or more Java files). However, this one integrates the gradual type system written in Stratego, takes into account that all error messages should be recomputed when the relevant file changes (be because error messages have positions in the file), and Pie works better (i.e. with less overhead) when you depend on (parts of) task outputs directly (the old design copied or referred to a lot of data from task output objects in other task inputs). The handling of modules and imports is also improved, using only the mb.resource vfs.

The package structure of the project has changed quite a bit to make things a bit more clearly separated and to be able to easily find things. TaskDefs live in a package task, input objects and output objects are each in a separate subpackage (task.input and task.output). This should allow of easy check if all input/output objects have an equals/hashcode specified for example. Similarly there is a packagefunction and function.output for serializable (projection) functions to select only part of a task output when depending only on that part of the task output.

The list of TaskDefs is as follows:

    Front: Asks for a module or library AST, extracts some ModuleData from that AST which will be used later.
    Resolve: Starts at the give "main" Stratego module, calls Front on it, and resolves the imports transitively from there to discover all modules. Then builds up some GlobalData from the ModuleData which is used in different places afterwards.
    CheckModule: Runs static checks on a module, based on the ModuleData and that of the modules are visible through imports. This is currently still transitive imports, but we may drop support of that to improve performance of the compiler. But that might influence other meta-languages that generate Stratego code. Static checks are mostly done in Stratego, although some overlap checks were easier to keep in Java.
    Check: Runs Resolve, gets the list of all modules, then runs CheckModule on each module. This task can be used to get all error messages for a project.
    Back: Runs per strategy definition. This task desugars the strategy, and then generates Java code for it. In the old architecture desugaring was done earlier in the process and cached in Pie, which is no longer necessary now. Late desugaring does means that dynamic rules are only lifted out at this point, which complicates things slightly. So strategies that contribute to the same dynamic rule are bundled into one larger Back task instead of being compiled separately. Desugaring output of these dynamic rule Back tasks are analysed as well, to be able to generate the DYNAMIC_CALLS strategy correctly.
    Another artefact of desugaring late is that we don't care to pass constructors to the desugaring. Therefore no congruence strategies are generated. Instead we do this in a separate Back task for all constructors at once. If this turns out to be costly we can split it up to generate per module or even per constructor.
    Compile: The one task to rule them all, this task runs Check, stops if there are errors, and otherwise continues to run all the Back tasks. It returns a list of the Java files that were written to, as well as non-error messages from Check.

# Migrating to this new code

If you used the old incremental compiler in a Spoofax project, that should continue to work as expected.
If you used the old incremental compiler from Java inside Spoofax 2, you will now need to use Pie TaskDef Compile or Check instead of StrIncr or StrIncrAnalysis respectively. See the changes in metaborg/spoofax#78 for examples of how this changed.
If you used the old incremental compiler from Java outside Spoofax 2, you also needed to implement the same interfaces as before: ParseStratego and ResourcePathConverter.