Finish work on always providing files from the same Pie task
Dynamic rules are now compiled in a separate task so the results can be queried by Boilerplate and Congruence Back tasks, instead of receiving that information from Compile. This was interfering with using a key for BackInput.

There are also some fixes to correctly identify resulting files that should be considered output files for a Back tasks (i.e. not collatoralStrategyOutput).