Fix bug in Stratego 2 compiler backend
Bug description: when some but not all strategy contributions defined a dynamic rule, the ones that didn't define the dynamic rule were ignored.

This commit makes sure strategy contributions in a BackInput.DynamicRule class are properly tracked, by not just looking for the set of related dynamic rules to compile together, but also _all_ the strategy contributions of the top-level strategy definitions that define the set of found dynamic rules. The adapted method now has 6 nested loops, gotta love that cyclomatic complexity metric.